requirements.txt
# VibeDoc Agentapplication 核心 dependency
# 针 toMCP&Agent挑战赛2025optimize的dependency包

# 🔧 核心 framework
gradio==5.34.1              # Agentinterface framework - 为用户提供直观的Agent交互体验
requests>=2.31.0            # HTTP请 requirement - 用于MCPservice通信
urllib3>=1.26.0
certifi>=2022.12.7

# 🤖 Agentconfiguration 管理  
python-dotenv>=0.19.0
pydantic>=1.10.0
typing-extensions>=4.4.0

# 📊 Agentasynchronous process
aiofiles>=22.0
anyio>=3.0

# 🚀 Agentapplication server
fastapi>=0.115.2
uvicorn>=0.14.0

# 📝 Agentcontent process
markdown>=3.8.2             # Markdownprocess - Agentgenerate contentformatting

# 📋 多 format 导出 support
python-docx>=1.2.0          # Worddocumentation 导出
reportlab>=4.4.3            # PDFgenerate support
html2text>=2024.4.24        # HTMLconvert

# Agent容器化 support (可选)
# weasyprint>=57.0  # 需 want 额外 system dependency
# zipfile36>=0.1.3  # Python 3.6+ 内置 support